Make the implementation of instructions more reusable
Since instructions of a specific category often perform the same logic,
such duplicates may reduce the maintainability and reusability.

To separate the registers used by an instruction from the logic
applied to them, I have to move the decoding logic back to
the scope where I'm able to access the registers and pass them by
reference as parameters.

Note that I'm not willing to shift the overhead of decoding from compile
time to runtime. I still possess a 1-to-1 mapping for almost 512 cases.

As for the logic of instructions, I decide to use functors rather than
lambdas so that I can reuse the code.