Refactor the structure of Registers
Make the PairedRegister store its internal data as high byte and low
byte separately. The low byte is a variant type which may contains
std::uint8_t or FlagRegister.

With the abovementioned adjustment, the accumulator (named as a) register
and the flag (named as f) can be stored within as well as passed as a
PairedRegister. Such improvement increases the expressiveness when using
interfaces taking these registers as parameters.