Make the boot process more fluent
The BootLoader should own the cartridge if there is one. Thus, it can
easily delegate the responsibility to the cartridge itself wherever
access to the specific range (0x0100 - 0x7FFF) of the address space
occurs.

The Cartridge should perform a similar workflow. It owns a Mbc object -
created during construction - for delegating the responsibility to it
where possible.